Fixed replace bug

+++ fastsend.py
@@ -70,13 +70,15 @@ def move_folders_and_copy_images(incoming_path, completed_path, processed_path, 
                 copy_images_to_completed(processed_folder_path, completed_parent_folder)
 
                 try:
-                    # Move the processed folder to another folder
-                    shutil.move(processed_folder_path, done_path)
+                    if os.path.exists(done_path):
+                        shutil.rmtree(done_path)  # Remove existing destination folder
+                        shutil.move(processed_folder_path, done_path)  # Move the processed folder
 
                     # Optionally, remove the prefix folder from the Completed folder
                     shutil.rmtree(os.path.join(completed_parent_folder, processed_folder), ignore_errors=True)
 
                     print("Program Executed Successfully...")
+                    
                 except Exception as e:
                     print(f"An error occurred: {e}")
 
@@ -107,7 +109,7 @@ def main():
                 json.dump(config, config_file)
         
         move_folders_and_copy_images(config["incoming_path"], config["completed_path"], config["processed_path"], config["done_path"])
-        time.sleep(60)
+        time.sleep(10)
 
 if __name__ == "__main__":
     main()
